handle predicates outside of core/state.StateDB
