+++ core/extstate/test_statedb.go
@@ -1,11 +1,12 @@
 // (c) 2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
-package state
+package extstate
 
 import (
 	"testing"
 
 	"github.com/ava-labs/coreth/core/rawdb"
+	"github.com/ava-labs/coreth/core/state"
 	"github.com/ava-labs/coreth/precompile/contract"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
@@ -13,7 +14,7 @@ import (
 
 func NewTestStateDB(t testing.TB) contract.StateDB {
 	db := rawdb.NewMemoryDatabase()
-	stateDB, err := New(common.Hash{}, NewDatabase(db), nil)
+	stateDB, err := state.New(common.Hash{}, state.NewDatabase(db), nil)
 	require.NoError(t, err)
-	return stateDB
+	return &StateDB{StateDB: stateDB}
 }
