+++ core/vm/interface.go
@@ -93,9 +93,6 @@ type StateDB interface {
 
 	AddLog(addr common.Address, topics []common.Hash, data []byte, blockNumber uint64)
 	GetLogData() (topics [][]common.Hash, data [][]byte)
-	GetPredicateStorageSlots(address common.Address, index int) ([]byte, bool)
-	SetPredicateStorageSlots(address common.Address, predicates [][]byte)
-
 	GetTxHash() common.Hash
 
 	AddPreimage(common.Hash, []byte)
